fix: refactor StoreBinding to remove unnecessary migration logic and improve store initialization
